MLOps-ImageClassification/requirements.txt
wget==3.2
azure-cognitiveservices-vision-customvision==3.0.0
python-decouple==3.3
azure-storage-file-share==12.2.0
flake8==3.8.4